Structure_function code
WAP to input name, roll number and marks in 5 subjects for n number of students. Write functions to -
a.Find total marks and percentage for all n students
b. Display details of a student with a given roll number
c.Display the details for all the students having % in a given range