Performing operations on 2d matrix
Let A be nXn square matrix. WAP by using appropriate user defined functions for the
following:
a) Find the number of nonzero elements in A
b) Find the sum of the elements above the leading diagonal.
c) Display the elements below the minor diagonal.
d) Find the product of the diagonal elements.